requirements.txt
fastapi
uvicorn
pydantic
httpx
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
loguru